+++ sdk/communication/azure-communication-identity/tests/test_communication_identity_client_async.py
@@ -33,8 +33,8 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             URIIdentityReplacer()])
     
     @CommunicationPreparer()
-    async def test_create_user_from_managed_identity(self, communication_connection_string):
-        endpoint, access_key = parse_connection_str(communication_connection_string)
+    async def test_create_user_from_managed_identity(self, communication_livetest_dynamic_connection_string):
+        endpoint, access_key = parse_connection_str(communication_livetest_dynamic_connection_string)
         from devtools_testutils import is_live
         if not is_live():
             credential = FakeTokenCredential()
@@ -51,9 +51,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert user.properties.get('id') is not None
 
     @CommunicationPreparer()
-    async def test_create_user(self, communication_connection_string):
+    async def test_create_user(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         async with identity_client:
@@ -62,9 +62,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert user.properties.get('id') is not None
 
     @CommunicationPreparer()
-    async def test_create_user_and_token(self, communication_connection_string):
+    async def test_create_user_and_token(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         async with identity_client:
@@ -74,8 +74,8 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert token_response.token is not None
 
     @CommunicationPreparer()
-    async def test_get_token_from_managed_identity(self, communication_connection_string):
-        endpoint, access_key = parse_connection_str(communication_connection_string)
+    async def test_get_token_from_managed_identity(self, communication_livetest_dynamic_connection_string):
+        endpoint, access_key = parse_connection_str(communication_livetest_dynamic_connection_string)
         from devtools_testutils import is_live
         if not is_live():
             credential = FakeTokenCredential()
@@ -94,9 +94,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert token_response.token is not None
 
     @CommunicationPreparer()
-    async def test_get_token(self, communication_connection_string):
+    async def test_get_token(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         async with identity_client:
@@ -107,8 +107,8 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert token_response.token is not None
 
     @CommunicationPreparer()
-    async def test_revoke_tokens_from_managed_identity(self, communication_connection_string):
-        endpoint, access_key = parse_connection_str(communication_connection_string)
+    async def test_revoke_tokens_from_managed_identity(self, communication_livetest_dynamic_connection_string):
+        endpoint, access_key = parse_connection_str(communication_livetest_dynamic_connection_string)
         from devtools_testutils import is_live
         if not is_live():
             credential = FakeTokenCredential()
@@ -128,9 +128,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert token_response.token is not None
 
     @CommunicationPreparer()
-    async def test_revoke_tokens(self, communication_connection_string):
+    async def test_revoke_tokens(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         async with identity_client:
@@ -142,8 +142,8 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert token_response.token is not None
 
     @CommunicationPreparer()
-    async def test_delete_user_from_managed_identity(self, communication_connection_string):
-        endpoint, access_key = parse_connection_str(communication_connection_string)
+    async def test_delete_user_from_managed_identity(self, communication_livetest_dynamic_connection_string):
+        endpoint, access_key = parse_connection_str(communication_livetest_dynamic_connection_string)
         from devtools_testutils import is_live
         if not is_live():
             credential = FakeTokenCredential()
@@ -161,9 +161,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert user.properties.get('id') is not None
 
     @CommunicationPreparer()
-    async def test_delete_user(self, communication_connection_string):
+    async def test_delete_user(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         async with identity_client:
@@ -173,9 +173,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         assert user.properties.get('id') is not None
     
     @CommunicationPreparer()
-    async def test_create_user_and_token_with_no_scopes(self, communication_connection_string):
+    async def test_create_user_and_token_with_no_scopes(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
 
@@ -184,9 +184,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
                 user, token_response = await identity_client.create_user_and_token(scopes=None)
 
     @CommunicationPreparer()
-    async def test_delete_user_with_no_user(self, communication_connection_string):
+    async def test_delete_user_with_no_user(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         
@@ -195,9 +195,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
                 await identity_client.delete_user(user=None)
 
     @CommunicationPreparer()
-    async def test_revoke_tokens_with_no_user(self, communication_connection_string):
+    async def test_revoke_tokens_with_no_user(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         
@@ -206,9 +206,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
                 await identity_client.revoke_tokens(user=None)
     
     @CommunicationPreparer()
-    async def test_get_token_with_no_user(self, communication_connection_string):
+    async def test_get_token_with_no_user(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )
         
@@ -217,9 +217,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
                 token_response = await identity_client.get_token(user=None, scopes=[CommunicationTokenScope.CHAT])
     
     @CommunicationPreparer()
-    async def test_get_token_with_no_scopes(self, communication_connection_string):
+    async def test_get_token_with_no_scopes(self, communication_livetest_dynamic_connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(
-            communication_connection_string,
+            communication_livetest_dynamic_connection_string,
             http_logging_policy=get_http_logging_policy()
         )