+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/aio/_async_container_repository_client.py
@@ -22,7 +22,7 @@ from .._helpers import _is_tag, _parse_next_link
 from .._models import (
     DeleteRepositoryResult,
     ContentProperties,
-    RegistryArtifactProperties,
+    ArtifactManifestProperties,
     RepositoryProperties,
     ArtifactTagProperties,
 )
@@ -45,6 +45,15 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :type credential: :class:`~azure.core.credentials_async.AsyncTokenCredential`
         :returns: None
         :raises: None
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/sample_create_client.py
+                :start-after: [START create_registry_client]
+                :end-before: [END create_registry_client]
+                :language: python
+                :dedent: 8
+                :caption: Instantiate an instance of `ContainerRepositoryClient`
         """
         if not endpoint.startswith("https://") and not endpoint.startswith("http://"):
             endpoint = "https://" + endpoint
@@ -64,6 +73,17 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :returns: Object containing information about the deleted repository
         :rtype: :class:`~azure.containerregistry.DeleteRepositoryResult`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            await client.delete()
         """
         return DeleteRepositoryResult._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.delete_repository(self.repository, **kwargs)
@@ -77,6 +97,18 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :type digest: str
         :returns: None
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for artifact in client.list_registry_artifacts():
+                await client.delete_registry_artifact(artifact.digest)
         """
         await self._client.container_registry.delete_manifest(self.repository, digest, **kwargs)
 
@@ -87,6 +119,18 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :param str tag: The tag to be deleted
         :returns: None
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for artifact in client.list_tags():
+                await client.delete_tag(tag.name)
         """
         await self._client.container_registry.delete_tag(self.repository, tag, **kwargs)
 
@@ -96,6 +140,24 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
 
         :returns: :class:`~azure.containerregistry.RepositoryProperties`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            repository_properties = await client.get_properties()
+            print(repository_properties.name)
+            print(repository_properties.content_permissions)
+            print(repository_properties.created_on)
+            print(repository_properties.last_updated_on)
+            print(repository_properties.manifest_count)
+            print(repository_properties.registry)
+            print(repository_properties.tag_count)
         """
         return RepositoryProperties._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.get_properties(self.repository, **kwargs)
@@ -104,19 +166,32 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
     @distributed_trace_async
     async def get_registry_artifact_properties(
         self, tag_or_digest: str, **kwargs: Dict[str, Any]
-    ) -> RegistryArtifactProperties:
+    ) -> ArtifactManifestProperties:
         """Get the properties of a registry artifact
 
         :param tag_or_digest: The tag/digest of a registry artifact
         :type tag_or_digest: str
-        :returns: :class:`~azure.containerregistry.RegistryArtifactProperties`
+        :returns: :class:`~azure.containerregistry.ArtifactManifestProperties`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for artifact in client.list_registry_artifacts():
+                properties = await client.get_registry_artifact_properties(artifact.digest)
         """
         if _is_tag(tag_or_digest):
             tag_or_digest = self._get_digest_from_tag(tag_or_digest)
 
-        return RegistryArtifactProperties._from_generated(  # pylint: disable=protected-access
-            await self._client.container_registry.get_manifest_properties(self.repository, tag_or_digest, **kwargs)
+        return ArtifactManifestProperties._from_generated(  # pylint: disable=protected-access
+            await self._client.container_registry.get_manifest_properties(self.repository, tag_or_digest, **kwargs),
+            repository=self.repository
         )
 
     @distributed_trace_async
@@ -127,6 +202,18 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :type tag: str
         :returns: :class:`~azure.containerregistry.ArtifactTagProperties`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for tag in client.list_tags():
+                tag_properties = await client.get_tag_properties(tag.name)
         """
         return ArtifactTagProperties._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.get_tag_properties(self.repository, tag, **kwargs),
@@ -134,19 +221,31 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         )
 
     @distributed_trace
-    def list_registry_artifacts(self, **kwargs: Dict[str, Any]) -> AsyncItemPaged[RegistryArtifactProperties]:
+    def list_registry_artifacts(self, **kwargs: Dict[str, Any]) -> AsyncItemPaged[ArtifactManifestProperties]:
         """List the artifacts for a repository
 
         :keyword last: Query parameter for the last item in the previous call. Ensuing
             call will return values after last lexically
         :paramtype last: str
         :keyword order_by: Query parameter for ordering by time ascending or descending
-        :paramtype order_by: :class:`~azure.containerregistry.RegistryArtifactOrderBy`
+        :paramtype order_by: :class:`~azure.containerregistry.ManifestOrderBy`
         :keyword results_per_page: Number of repositories to return per page
         :paramtype results_per_page: int
-        :return: ItemPaged[:class:`~azure.containerregistry.RegistryArtifactProperties`]
+        :return: ItemPaged[:class:`~azure.containerregistry.ArtifactManifestProperties`]
         :rtype: :class:`~azure.core.async_paging.AsyncItemPaged`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for artifact in client.list_registry_artifacts():
+                print(artifact.digest)
         """
         name = self.repository
         last = kwargs.pop("last", None)
@@ -155,7 +254,7 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         cls = kwargs.pop(
             "cls",
             lambda objs: [
-                RegistryArtifactProperties._from_generated(x) for x in objs  # pylint: disable=protected-access
+                ArtifactManifestProperties._from_generated(x, repository=self.repository) for x in objs  # pylint: disable=protected-access
             ],
         )
 
@@ -264,6 +363,18 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :return: ItemPaged[:class:`~azure.containerregistry.ArtifactTagProperties`]
         :rtype: :class:`~azure.core.async_paging.AsyncItemPaged`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for tag in client.list_tags():
+                tag_properties = await client.get_tag_properties(tag.name)
         """
         name = self.repository
         last = kwargs.pop("last", None)
@@ -381,13 +492,34 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :type digest: str
         :param permissions: The property's values to be set
         :type permissions: ContentProperties
-        :returns: :class:`~azure.containerregistry.RegistryArtifactProperties`
+        :returns: :class:`~azure.containerregistry.ArtifactManifestProperties`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            async for artifact in client.list_registry_artifacts():
+                received_permissions = await client.set_manifest_properties(
+                    artifact.digest,
+                    ContentPermissions(
+                        can_delete=False,
+                        can_list=False,
+                        can_read=False,
+                        can_write=False,
+                    ),
+                )
         """
-        return RegistryArtifactProperties._from_generated(  # pylint: disable=protected-access
+        return ArtifactManifestProperties._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.update_manifest_properties(
                 self.repository, digest, value=permissions._to_generated(), **kwargs  # pylint: disable=protected-access
-            )
+            ),
+            repository=self.repository
         )
 
     @distributed_trace_async
@@ -402,6 +534,26 @@ class ContainerRepositoryClient(ContainerRegistryBaseClient):
         :type permissions: ContentProperties
         :returns: :class:`~azure.containerregistry.ArtifactTagProperties`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+
+        Example
+
+        .. code-block:: python
+
+            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.identity.aio import DefaultAzureCredential
+
+            account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
+            client = ContainerRepositoryClient(account_url, "my_repository", DefaultAzureCredential())
+            tag_identifier = "latest"
+            received = await client.set_tag_properties(
+                tag_identifier,
+                ContentPermissions(
+                    can_delete=False,
+                    can_list=False,
+                    can_read=False,
+                    can_write=False,
+                ),
+            )
         """
         return ArtifactTagProperties._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.update_tag_attributes(