+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_generated/models/__init__.py
@@ -30,6 +30,9 @@ try:
     from ._models_py3 import AcsChatThreadPropertiesUpdatedEventData
     from ._models_py3 import AcsChatThreadPropertiesUpdatedPerUserEventData
     from ._models_py3 import AcsChatThreadWithUserDeletedEventData
+    from ._models_py3 import AcsRecordingChunkInfoProperties
+    from ._models_py3 import AcsRecordingFileStatusUpdatedEventData
+    from ._models_py3 import AcsRecordingStorageInfoProperties
     from ._models_py3 import AcsSmsDeliveryAttemptProperties
     from ._models_py3 import AcsSmsDeliveryReportReceivedEventData
     from ._models_py3 import AcsSmsEventBaseProperties
@@ -137,9 +140,11 @@ try:
     from ._models_py3 import ServiceBusDeadletterMessagesAvailableWithNoListenersEventData
     from ._models_py3 import SignalRServiceClientConnectionConnectedEventData
     from ._models_py3 import SignalRServiceClientConnectionDisconnectedEventData
+    from ._models_py3 import StorageAsyncOperationInitiatedEventData
     from ._models_py3 import StorageBlobCreatedEventData
     from ._models_py3 import StorageBlobDeletedEventData
     from ._models_py3 import StorageBlobRenamedEventData
+    from ._models_py3 import StorageBlobTierChangedEventData
     from ._models_py3 import StorageDirectoryCreatedEventData
     from ._models_py3 import StorageDirectoryDeletedEventData
     from ._models_py3 import StorageDirectoryRenamedEventData
@@ -186,6 +191,9 @@ except (SyntaxError, ImportError):
     from ._models import AcsChatThreadPropertiesUpdatedEventData  # type: ignore
     from ._models import AcsChatThreadPropertiesUpdatedPerUserEventData  # type: ignore
     from ._models import AcsChatThreadWithUserDeletedEventData  # type: ignore
+    from ._models import AcsRecordingChunkInfoProperties  # type: ignore
+    from ._models import AcsRecordingFileStatusUpdatedEventData  # type: ignore
+    from ._models import AcsRecordingStorageInfoProperties  # type: ignore
     from ._models import AcsSmsDeliveryAttemptProperties  # type: ignore
     from ._models import AcsSmsDeliveryReportReceivedEventData  # type: ignore
     from ._models import AcsSmsEventBaseProperties  # type: ignore
@@ -293,9 +301,11 @@ except (SyntaxError, ImportError):
     from ._models import ServiceBusDeadletterMessagesAvailableWithNoListenersEventData  # type: ignore
     from ._models import SignalRServiceClientConnectionConnectedEventData  # type: ignore
     from ._models import SignalRServiceClientConnectionDisconnectedEventData  # type: ignore
+    from ._models import StorageAsyncOperationInitiatedEventData  # type: ignore
     from ._models import StorageBlobCreatedEventData  # type: ignore
     from ._models import StorageBlobDeletedEventData  # type: ignore
     from ._models import StorageBlobRenamedEventData  # type: ignore
+    from ._models import StorageBlobTierChangedEventData  # type: ignore
     from ._models import StorageDirectoryCreatedEventData  # type: ignore
     from ._models import StorageDirectoryDeletedEventData  # type: ignore
     from ._models import StorageDirectoryRenamedEventData  # type: ignore
@@ -355,6 +365,9 @@ __all__ = [
     'AcsChatThreadPropertiesUpdatedEventData',
     'AcsChatThreadPropertiesUpdatedPerUserEventData',
     'AcsChatThreadWithUserDeletedEventData',
+    'AcsRecordingChunkInfoProperties',
+    'AcsRecordingFileStatusUpdatedEventData',
+    'AcsRecordingStorageInfoProperties',
     'AcsSmsDeliveryAttemptProperties',
     'AcsSmsDeliveryReportReceivedEventData',
     'AcsSmsEventBaseProperties',
@@ -462,9 +475,11 @@ __all__ = [
     'ServiceBusDeadletterMessagesAvailableWithNoListenersEventData',
     'SignalRServiceClientConnectionConnectedEventData',
     'SignalRServiceClientConnectionDisconnectedEventData',
+    'StorageAsyncOperationInitiatedEventData',
     'StorageBlobCreatedEventData',
     'StorageBlobDeletedEventData',
     'StorageBlobRenamedEventData',
+    'StorageBlobTierChangedEventData',
     'StorageDirectoryCreatedEventData',
     'StorageDirectoryDeletedEventData',
     'StorageDirectoryRenamedEventData',