+++ sdk/tables/azure-data-tables/azure/data/tables/_table_client.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 
 import functools
-from typing import Optional, Any, Union, List, Tuple, Dict, Mapping, Iterable
+from typing import Optional, Any, Union, List, Tuple, Dict, Mapping, Iterable, overload
 try:
     from urllib.parse import urlparse, unquote
 except ImportError:
@@ -22,6 +22,7 @@ from ._error import _process_table_error, _validate_table_name
 from ._generated.models import (
     SignedIdentifier,
     TableProperties,
+    QueryOptions
 )
 from ._serialize import _get_match_headers, _add_entity_properties
 from ._base_client import parse_connection_str, TablesBaseClient
@@ -41,40 +42,40 @@ TransactionOperationType = Union[Tuple[OperationType, EntityType], Tuple[Operati
 
 
 class TableClient(TablesBaseClient):
-    """
-    :ivar str account_name: Name of the storage account (Cosmos or Azure)
-    :ivar str table_name: The name of the table
+    """A client to interact with a specific Table in an Azure Tables account.
+
+    :ivar str account_name: The name of the Tables account.
+    :ivar str table_name: The name of the table.
+    :ivar str url: The full URL to the Tables account.
     """
 
     def __init__(
         self,
-        account_url,  # type: str
+        endpoint,  # type: str
         table_name,  # type: str
-        credential=None,  # type: str
+        credential=None,  # type: Union[AzureNamedKeyCredential, AzureSasCredential]
         **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Create TableClient from a Credential.
 
-        :param account_url:
-            A url to an Azure Storage account.
-        :type account_url: str
-        :param table_name: The table name.
-        :type table_name: str
+        :param str endpoint: A URL to an Azure Tables account.
+        :param str table_name: The table name.
         :param credential:
             The credentials with which to authenticate. This is optional if the
             account URL already has a SAS token, or the connection string already has shared
             access key values. The value can be a SAS token string or an account shared access
             key.
-        :type credential: str
-
+        :type credential:
+            :class:`~azure.core.credentials.AzureNamedKeyCredential` or
+            :class:`~azure.core.credentials.AzureSasCredential`
         :returns: None
         """
         if not table_name:
             raise ValueError("Please specify a table name.")
         _validate_table_name(table_name)
         self.table_name = table_name
-        super(TableClient, self).__init__(account_url, credential=credential, **kwargs)
+        super(TableClient, self).__init__(endpoint, credential=credential, **kwargs)
 
     def _format_url(self, hostname):
         """Format the endpoint URL according to the current location
@@ -92,11 +93,8 @@ class TableClient(TablesBaseClient):
         # type: (...) -> TableClient
         """Create TableClient from a Connection String.
 
-        :param conn_str:
-            A connection string to an Azure Storage or Cosmos account.
-        :type conn_str: str
-        :param table_name: The table name.
-        :type table_name: str
+        :param str conn_str: A connection string to an Azure Tables account.
+        :param str table_name: The table name.
         :returns: A table client.
         :rtype: :class:`~azure.data.tables.TableClient`
 
@@ -109,10 +107,10 @@ class TableClient(TablesBaseClient):
                 :dedent: 8
                 :caption: Authenticating a TableServiceClient from a connection_string
         """
-        account_url, credential = parse_connection_str(
+        endpoint, credential = parse_connection_str(
             conn_str=conn_str, credential=None, keyword_args=kwargs
         )
-        return cls(account_url, table_name=table_name, credential=credential, **kwargs)
+        return cls(endpoint, table_name=table_name, credential=credential, **kwargs)
 
     @classmethod
     def from_table_url(cls, table_url, credential=None, **kwargs):
@@ -142,7 +140,7 @@ class TableClient(TablesBaseClient):
         account_path = ""
         if len(table_path) > 1:
             account_path = "/" + "/".join(table_path[:-1])
-        account_url = "{}://{}{}?{}".format(
+        endpoint = "{}://{}{}?{}".format(
             parsed_url.scheme,
             parsed_url.netloc.rstrip("/"),
             account_path,
@@ -153,7 +151,7 @@ class TableClient(TablesBaseClient):
             raise ValueError(
                 "Invalid URL. Please provide a URL with a valid table name"
             )
-        return cls(account_url, table_name=table_name, credential=credential, **kwargs)
+        return cls(endpoint, table_name=table_name, credential=credential, **kwargs)
 
     @distributed_trace
     def get_table_access_policy(
@@ -253,11 +251,12 @@ class TableClient(TablesBaseClient):
         self, **kwargs  # type: Any
     ):
         # type: (...) -> None
-        """Deletes the table under the current account.
+        """Deletes the table under the current account. No error will be raised
+            if the table does not exist
 
         :return: None
         :rtype: None
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError` If the table does not exist
+        :raises: :class:`~azure.core.exceptions.HttpResponseError`
 
         .. admonition:: Example:
 
@@ -271,17 +270,25 @@ class TableClient(TablesBaseClient):
         try:
             self._client.table.delete(table=self.table_name, **kwargs)
         except HttpResponseError as error:
+            if error.status_code == 404:
+                return
             _process_table_error(error)
 
+    @overload
+    def delete_entity(self, partition_key, row_key, **kwargs):
+        # type: (str, str, Any) -> None
+        pass
+
+    @overload
+    def delete_entity(self, entity, **kwargs):
+        # type: (Union[TableEntity, Mapping[str, Any]], Any) -> None
+        pass
+
     @distributed_trace
-    def delete_entity(
-        self,
-        partition_key,  # type: str
-        row_key,  # type: str
-        **kwargs  # type: Any
-    ):
-        # type: (...) -> None
-        """Deletes the specified entity in a table.
+    def delete_entity(self, *args, **kwargs):
+        # type: (Union[TableEntity, str], Any) -> None
+        """Deletes the specified entity in a table. No error will be raised if
+            the entity or PartitionKey-RowKey pairing is not found.
 
         :param partition_key: The partition key of the entity.
         :type partition_key: str
@@ -292,7 +299,7 @@ class TableClient(TablesBaseClient):
         :paramtype match_condition: ~azure.core.MatchConditions
         :return: None
         :rtype: None
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError` If the entity already does not exist
+        :raises: :class:`~azure.core.exceptions.HttpResponseError`
 
         .. admonition:: Example:
 
@@ -303,12 +310,33 @@ class TableClient(TablesBaseClient):
                 :dedent: 8
                 :caption: Deleting an entity to a Table
         """
+        try:
+            entity = kwargs.pop('entity', None)
+            if not entity:
+                entity = args[0]
+            partition_key = entity['PartitionKey']
+            row_key = entity['RowKey']
+        except (TypeError, IndexError):
+            partition_key = kwargs.pop('partition_key', None)
+            if not partition_key:
+                partition_key = args[0]
+            row_key = kwargs.pop("row_key", None)
+            if not row_key:
+                row_key = args[1]
+
+        match_condition = kwargs.pop("match_condition", None)
+        etag = kwargs.pop("etag", None)
+        if match_condition and entity and not etag:
+            try:
+                etag = entity.metadata.get("etag", None)
+            except (AttributeError, TypeError):
+                pass
 
         if_match, _ = _get_match_headers(
             kwargs=dict(
                 kwargs,
-                etag=kwargs.pop("etag", None),
-                match_condition=kwargs.pop("match_condition", None),
+                etag=etag,
+                match_condition=match_condition,
             ),
             etag_param="etag",
             match_param="match_condition",
@@ -323,6 +351,8 @@ class TableClient(TablesBaseClient):
                 **kwargs
             )
         except HttpResponseError as error:
+            if error.status_code == 404:
+                return
             _process_table_error(error)
 
     @distributed_trace
@@ -338,7 +368,7 @@ class TableClient(TablesBaseClient):
         :type entity: :class:`~azure.data.tables.TableEntity` or Dict[str,str]
         :return: Dictionary mapping operation metadata returned from the service
         :rtype: Dict[str,str]
-        :raises: :class:`~azure.core.exceptions.ResourceExistsError` If the entity already exists
+        :raises: :class:`~azure.core.exceptions.HttpResponseError`
 
         .. admonition:: Example:
 
@@ -394,12 +424,19 @@ class TableClient(TablesBaseClient):
                 :dedent: 8
                 :caption: Updating an already exiting entity in a Table
         """
+        match_condition = kwargs.pop("match_condition", None)
+        etag = kwargs.pop("etag", None)
+        if match_condition and not etag:
+            try:
+                etag = entity.metadata.get("etag", None)
+            except (AttributeError, TypeError):
+                pass
 
         if_match, _ = _get_match_headers(
             kwargs=dict(
                 kwargs,
-                etag=kwargs.pop("etag", None),
-                match_condition=kwargs.pop("match_condition", None),
+                etag=etag,
+                match_condition=match_condition,
             ),
             etag_param="etag",
             match_param="match_condition",
@@ -443,8 +480,8 @@ class TableClient(TablesBaseClient):
         # type: (...) -> ItemPaged[TableEntity]
         """Lists entities in a table.
 
-        :keyword int results_per_page: Number of entities per page in return ItemPaged
-        :keyword select: Specify desired properties of an entity to return certain entities
+        :keyword int results_per_page: Number of entities returned per service request.
+        :keyword select: Specify desired properties of an entity to return.
         :paramtype select: str or List[str]
         :return: ItemPaged[:class:`~azure.data.tables.TableEntity`]
         :rtype: ~azure.core.paging.ItemPaged
@@ -483,8 +520,8 @@ class TableClient(TablesBaseClient):
         """Lists entities in a table.
 
         :param str query_filter: Specify a filter to return certain entities
-        :keyword int results_per_page: Number of entities per page in return ItemPaged
-        :keyword select: Specify desired properties of an entity to return certain entities
+        :keyword int results_per_page: Number of entities returned per service request.
+        :keyword select: Specify desired properties of an entity to return.
         :paramtype select: str or List[str]
         :keyword Dict[str, Any] parameters: Dictionary for formatting query with additional, user defined parameters
         :return: ItemPaged[:class:`~azure.data.tables.TableEntity`]
@@ -507,7 +544,7 @@ class TableClient(TablesBaseClient):
         top = kwargs.pop("results_per_page", None)
         user_select = kwargs.pop("select", None)
         if user_select and not isinstance(user_select, str):
-            user_select = ", ".join(user_select)
+            user_select = ",".join(user_select)
 
         command = functools.partial(self._client.table.query_entities, **kwargs)
         return ItemPaged(
@@ -533,6 +570,8 @@ class TableClient(TablesBaseClient):
         :type partition_key: str
         :param row_key: The row key of the entity.
         :type row_key: str
+        :keyword select: Specify desired properties of an entity to return.
+        :paramtype select: str or List[str]
         :return: Dictionary mapping operation metadata returned from the service
         :rtype: :class:`~azure.data.tables.TableEntity`
         :raises: :class:`~azure.core.exceptions.HttpResponseError`
@@ -546,14 +585,17 @@ class TableClient(TablesBaseClient):
                 :dedent: 8
                 :caption: Get a single entity from a table
         """
+        user_select = kwargs.pop("select", None)
+        if user_select and not isinstance(user_select, str):
+            user_select = ",".join(user_select)
         try:
             entity = self._client.table.query_entity_with_partition_and_row_key(
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
+                query_options=QueryOptions(select=user_select),
                 **kwargs
             )
-
             properties = _convert_to_entity(entity)
             return properties
         except HttpResponseError as error: