+++ sdk/tables/azure-data-tables/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 12.0.0b7 (2021-05-11)
 **Breaking**
+* The `account_url` parameter in the client constructors has been renamed to `endpoint`.
+* The `TableEntity` object now acts exclusively like a dictionary, and no longer supports key access via attributes.
+* Metadata of an entity is now accessed via `TableEntity.metadata` attribute rather than a method.
 * Removed explicit `LinearRetry` and `ExponentialRetry` in favor of keyword parameter.
 * Renamed `filter` parameter in query APIs to `query_filter`.
 * The `location_mode` attribute on clients is now read-only. This has been added as a keyword parameter to the constructor.
@@ -10,10 +13,11 @@
 * `TableClient.send_batch` has been renamed to `TableClient.submit_transaction`.
 * Removed `BatchTransactionResult` object in favor of returning an iterable of batched entities with returned metadata.
 * Removed Batching context-manager behavior
-* Changed optional `value` and `type` arguments of `EntityProperty` to required.
+* `EntityProperty` is now a NampedTuple, and can be represented by a tuple of `(entity, EdmType)`.
 * Renamed `EntityProperty.type` to `EntityProperty.edm_type`.
 * `BatchErrorException` has been renamed to `TableTransactionError`.
-* `EntityProperty` is now a tuple.
+* The `location_mode` is no longer a public attribute on the Clients.
+* The only supported credentials are `AzureNamedKeyCredential`, `AzureSasCredential`, or authentication by connection string
 * Removed `date` and `api_version` from the `TableItem` class.
 
 **Fixes**
@@ -24,6 +28,8 @@
 * Added support for Azurite storage emulator
 * Throws a `RequestTooLargeError` on transaction requests that return a 413 error code
 * Added support for Int64 and Binary types in query filters
+* Added support for `select` keyword parameter to `TableClient.get_entity()`.
+* On `update_entity` and `delete_entity` if no `etag` is supplied via kwargs, the `etag` in the entity will be used if it is in the entity.
 
 ## 12.0.0b6 (2021-04-06)
 * Updated deserialization of datetime fields in entities to support preservation of the service format with additional decimal place.